removing program class
