+++ Assets/Mirror/Editor/Weaver/Log.cs
@@ -1,0 +1,20 @@
+using System;
+
+namespace Mirror.Weaver
+{
+    public static class Log
+    {
+        public static Action<string> WarningMethod;
+        public static Action<string> ErrorMethod;
+
+        public static void Warning(string msg)
+        {
+            WarningMethod(msg);
+        }
+
+        public static void Error(string msg)
+        {
+            ErrorMethod(msg);
+        }
+    }
+}
